Added showAmountsOwed to UserManager.java and updated Person to contain the Hashmap amountsOwed. Changed the package names to be in line with naming conventions. Updated GroupManager.java to update the amountsOwed hashmap on payment.
