+++ README.md
@@ -1,17 +1,20 @@
 # Group 54
+
+## Prerequisites
+- Please ensure you have jdk11. Use Amazon Coretto for maximum compatibility.
+- Please ensure you have JUnit Tests added to class path and junit.jupiter
+- IntelliJ should install it on highlight in the Unit Tests.
+
 ## Members
 - [Rohan Tinna](https://rohantinna.com) (rohan.tinna@mail.utoronto.ca)
 - Subhasis Chakraborti (subhasis.chakraborti@mail.utoronto.ca)
 - Lingyun Jiang (johnnyl.jiang@mail.utoronto.ca)
 - Wasee Alam (wasee.alam@mail.utoronto.ca)
-- Add your details by editing `README.md`
+- Mohamed Al-Fahim (m.alfahim@mail.utoronto.ca)
 
 ## Links
 - [Project Ideas](https://docs.google.com/document/d/1VQAk6e_4F3ED4pP00Pl9-hLnjafwfLhiBB_jGU9WbmA/edit#heading=h.cbvciyafl8gz)
 - [Discord Channel](https://discord.gg/QgV5UTxfhp)
 - [Course Notes](https://uoft.rohantinna.in/notes/csc207)
-- Specification
-- [CRC Cards](https://utoronto-my.sharepoint.com/:o:/g/personal/rohan_tinna_mail_utoronto_ca/Ettb3QDSfqVJtKyJhTRb3FEBf6VCMOUHe8U84aeBMTvf3Q?e=8lrg3s)
-
-## Important Dates & Deadlines
-- `Phase 0` due Friday, Oct 15 (Saturday with extension)
+- [Specification](https://docs.google.com/document/d/1VQAk6e_4F3ED4pP00Pl9-hLnjafwfLhiBB_jGU9WbmA/edit#heading=h.cbvciyafl8gz)
+- [CRC Cards](https://utoronto-my.sharepoint.com/:p:/g/personal/wasee_alam_mail_utoronto_ca/EU4orkY_c35LjHMskrwIRtUBlcrT7rPn1OHK67QFkBEDwQ?e=7shx5g)
