--- src/main/java/Use_Cases/UserManager.java
@@ -1,17 +1,0 @@
-/*
-Below is the UserManager class which edits and returns instances of User.
- */
-package Use_Cases;
-
-import entities.*;
-
-public class UserManager {
-
-    public static double showBalance(Person p){
-        return p.getBalance();
-    }
-
-    public static void updateProfile(Person p) {
-        //TODO: Finish this
-    }
-}
