+++ src/main/java/entities/Person.java
@@ -6,10 +6,13 @@
 
 package entities;
 
+import java.util.HashMap;
+
 public class Person {
     private String name;
     private double balance;
     private String email;
+    private HashMap<Group, Double> amountsOwed;
 
     /**
      * Construct Person, giving them the given name, balance, and email.
@@ -21,12 +24,14 @@ public class Person {
         this.name = name;
         this.balance = balance;
         this.email = email;
+        this.amountsOwed = null;
     }
 
     public Person(double balance, String email){
         this.name = email;
         this.balance = balance;
         this.email = email;
+        this.amountsOwed = null;
     }
 
     public String getName() {
@@ -53,4 +58,13 @@ public class Person {
         this.email = email;
     }
 
+    public HashMap<Group, Double> getAmountsOwed() {
+        return amountsOwed;
+    }
+
+    public void setAmountsOwed(Group group, double owing) {
+        HashMap<Group, Double> tempHash = new HashMap<>();
+        tempHash.put(group, owing);
+        this.amountsOwed = tempHash;
+    }
 }
